Colocate tuners and oracles.
